Move doc migration stuff into its own file
