+++ packages/browser-extension/src/sync/sync.ts
@@ -10,12 +10,7 @@ import {
   setSyncServerAddresses,
 } from "./syncServerStorage";
 import { getRepo } from "./repos";
-import {
-  MapMigrationIndex,
-  mapMigrations,
-  persistedStateVersion,
-} from "../store/migrations";
-import * as appLogger from "../logging/appLogging";
+import { triggerMigrationIfNecessary } from "./migrations";
 
 export function createDoc(map: NewArgumentMap) {
   const repo = getRepo([]);
@@ -90,31 +85,4 @@ function getRepoForDoc(id: DocumentId) {
 export interface NewArgumentMap
   extends Omit<ArgumentMap, "automergeDocumentId"> {
   automergeDocumentId?: string;
-}
-
-// The last version before we switched to Automerge. Automerge maps missing a version
-// are implied to have this version.
-const minAutomergeMapVersion = 8;
-
-export function triggerMigrationIfNecessary(handle: DocHandle<ArgumentMap>) {
-  ensureMapMigrations(handle).catch((reason) =>
-    appLogger.error("Failed to migrate doc", reason),
-  );
-}
-
-async function ensureMapMigrations(handle: DocHandle<ArgumentMap>) {
-  const doc = await handle.doc();
-  if (!doc) {
-    throw new Error("Unable to get doc for migration");
-  }
-  let currentVersion = doc.version || minAutomergeMapVersion;
-  if (currentVersion < persistedStateVersion) {
-    handle.change((map) => {
-      while (currentVersion <= persistedStateVersion) {
-        mapMigrations[currentVersion as MapMigrationIndex]?.(map);
-        map.version = currentVersion;
-        currentVersion++;
-      }
-    });
-  }
 }
